Increase precision for conversions
This prevents some rounding problems for multiplication results,
as 16 bit posits exceed (in multiplication) the precision of float32,
and likewise 32 bit posits exceed the precision of float64.

+++ src/codec.h
@@ -19,8 +19,8 @@ long double codec_linear_l_from_float64(double);
 long double codec_linear_l_from_extended_float(long double);
 
 posit_log8 codec_posit_log8_from_s_and_l(bool, float);
-posit_log16 codec_posit_log16_from_s_and_l(bool, float);
-posit_log32 codec_posit_log32_from_s_and_l(bool, double);
+posit_log16 codec_posit_log16_from_s_and_l(bool, double);
+posit_log32 codec_posit_log32_from_s_and_l(bool, long double);
 posit_log64 codec_posit_log64_from_s_and_l(bool, long double);
 
 #endif /* CODEC_H */
